+++ SoureCodeCommonBundle.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace SoureCode\Bundle\Common;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class SoureCodeCommonBundle extends Bundle
+{
+}